requirements.txt
Flask==2.2.2
Jinja2==3.1.2
flask-mysqldb==1.0.1
pyodbc==4.0.39
mysql-connector-python=8.1.0
